requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
python-dotenv==1.0.1
customtkinter==5.2.2
pyperclip==1.9.0
psutil==6.1.0
pyautogui==0.9.54
mss==10.0.0
screen-brightness-control==0.24.1
winshell==0.6
pywin32==308
py-cpuinfo==9.0.0
speedtest-cli==2.1.3
Pillow==11.0.0
